+++ imports/api/labels/labels.js
@@ -3,15 +3,18 @@ import { Mongo } from 'meteor/mongo';
 import { check } from 'meteor/check';
 import { Tasks } from "/imports/api/tasks/tasks";
 import { checkLoggedIn, checkCanReadProject, checkCanWriteProject } from "/imports/api/permissions/permissions"
+import LabelSchema from './schema';
 
 export const Labels = new Mongo.Collection('labels');
+Labels.attachSchema(LabelSchema);
+Labels.methods = {}
+
 if (Meteor.isServer) {
   Meteor.startup(() => {
     Tasks.rawCollection().createIndex({projectId: 1});
   });
 }
 
-Labels.methods = {}
 
 Labels.methods.create = new ValidatedMethod({
   name: "labels.create",