+++ imports/api/projectGroups/projectGroups.js
@@ -1,8 +1,11 @@
 import { Meteor } from 'meteor/meteor';
 import { Mongo } from 'meteor/mongo';
 import { check } from 'meteor/check';
+import ProjectGroupSchema from "./schema";
 
 export const ProjectGroups = new Mongo.Collection('projectGroups');
+ProjectGroups.attachSchema(ProjectGroupSchema);
+
 if (Meteor.isServer) {
   Meteor.startup(() => {
     ProjectGroups.rawCollection().createIndex({organizationId: 1});