+++ imports/api/projects/projects.js
@@ -11,8 +11,12 @@ import { Canvas } from "/imports/api/canvas/canvas.js";
 import { Labels } from "/imports/api/labels/labels.js";
 import { Events } from "/imports/api/events/events.js";
 import { Permissions, checkLoggedIn, checkCanReadProject, checkCanWriteProject } from "/imports/api/permissions/permissions"
+import ProjectSchema from './schema';
 
 export const Projects = new Mongo.Collection("projects");
+Projects.attachSchema(ProjectSchema);
+Projects.methods = {};
+
 if (Meteor.isServer) {
   Meteor.startup(() => {
     Projects.rawCollection().createIndex({ organizationId: 1 });
@@ -32,7 +36,6 @@ export const ProjectAccessRights = Object.freeze({
   PRIVATE: "private",
 });
 
-Projects.methods = {};
 
 const checkIfAdminOrCreator = (projectId) => {
   if (Permissions.isAdmin(Meteor.userId())) {
@@ -325,7 +328,6 @@ Projects.methods.clone = new ValidatedMethod({
       startDate: project.startDate,
       members: project.members,
       endDate: project.endDate,
-      estimatedSize: project.estimatedSize,
       color: project.color,
       features: project.features
     });
@@ -488,22 +490,6 @@ Projects.methods.setDatesAndState = new ValidatedMethod({
     checkLoggedIn();
     checkIfAdminOrCreator(projectId);
     Projects.update({ _id: projectId }, { $set: { startDate: startDate, endDate: endDate, state: state } });
-  }
-});
-
-Projects.methods.updateEstimatedSize = new ValidatedMethod({
-  name: "projects.updateEstimatedSize",
-  validate: new SimpleSchema({
-    projectId: { type: String },
-    estimatedSize: { type: Number }
-  }).validator(),
-  run({projectId, estimatedSize}) {
-    checkLoggedIn();
-    checkIfAdminOrCreator(projectId);
-    Projects.update(
-      { _id: projectId },
-      { $set: { estimatedSize: estimatedSize } }
-    );
   }
 });