+++ test/main.js
@@ -3,9 +3,12 @@ import "/server/main"
 import "/test/api/users/users.test"
 import "/test/api/bpmn/processDiagrams.test"
 import "/test/api/projects/projects.test"
+import "/test/api/tasks/tasks.test"
 import "/test/api/organizations/organizations.test"
 import "/test/api/labels/labels.test"
 import "/test/api/notifications/notifications.test"
+import "/test/api/canvas/canvas.test"
+import "/test/api/healthReports/healthReports.test"
 
 describe("latelier", function () {
   it("package.json has correct name", async function () {